requirements.txt
tk
pyodbc
pytest
customtkinter
Pillow
